- type traits go in type_traits.h - things found in <utility> go in utility.h - algs.h doesn't contain any type traits. It just includes type_traits.h
