+++ dlib/utility.h
@@ -9,7 +9,7 @@
 /*
     This header contains back-ports of C++14/17 functions and type traits
     found in <utility> header of the standard library.
- */
+*/
 
 namespace dlib
 {
@@ -51,90 +51,6 @@ namespace dlib
     using index_sequence_for = make_index_sequence<sizeof...(Ts)>;
     // ---------------------------------------------------------------------
 #endif
-
-    // ---------------------------------------------------------------------
-
-    template<bool First, bool... Rest>
-    struct And : std::integral_constant<bool, First && And<Rest...>::value> {};
-
-    template<bool Value>
-    struct And<Value> : std::integral_constant<bool, Value>{};
-
-    // ---------------------------------------------------------------------
-
-    template<class...>
-    struct conjunction : std::true_type {};
-
-    template<class B1>
-    struct conjunction<B1> : B1 {};
-
-    template<class B1, class... Bn>
-    struct conjunction<B1, Bn...> : std::conditional_t<bool(B1::value), conjunction<Bn...>, B1> {};
-
-    // ---------------------------------------------------------------------
-
-    template <typename ...Types>
-    struct are_nothrow_move_constructible : And<std::is_nothrow_move_constructible<Types>::value...> {};
-
-    // ---------------------------------------------------------------------
-
-    template <typename ...Types>
-    struct are_nothrow_move_assignable : And<std::is_nothrow_move_assignable<Types>::value...> {};
-
-    // ---------------------------------------------------------------------
-
-    template <typename ...Types>
-    struct are_nothrow_copy_constructible : And<std::is_nothrow_copy_constructible<Types>::value...> {};
-
-    // ---------------------------------------------------------------------
-
-    template <typename ...Types>
-    struct are_nothrow_copy_assignable : And<std::is_nothrow_copy_assignable<Types>::value...> {};
-
-    // ---------------------------------------------------------------------
-
-    template< class... >
-    using void_t = void;
-
-    // ---------------------------------------------------------------------
-
-    namespace swappable_details
-    {
-        using std::swap;
-
-        template<typename T, typename = void>
-        struct is_swappable : std::false_type {};
-
-        template<typename T>
-        struct is_swappable<T, void_t<decltype(swap(std::declval<T&>(), std::declval<T&>()))>> : std::true_type {};
-
-        template<typename T>
-        struct is_nothrow_swappable :
-            std::integral_constant<bool, is_swappable<T>::value &&
-                                         noexcept(swap(std::declval<T&>(), std::declval<T&>()))> {};
-    }
-
-    // ---------------------------------------------------------------------
-
-    template<typename T>
-    struct is_swappable : swappable_details::is_swappable<T>{};
-
-    // ---------------------------------------------------------------------
-
-    template<typename T>
-    struct is_nothrow_swappable : swappable_details::is_nothrow_swappable<T>{};
-
-    // ---------------------------------------------------------------------
-
-    template <typename ...Types>
-    struct are_nothrow_swappable : And<is_nothrow_swappable<Types>::value...> {};
-
-    // ---------------------------------------------------------------------
-
-    template<std::size_t I>
-    using size_ = std::integral_constant<std::size_t, I>;
-
-    // ---------------------------------------------------------------------
 }
 
 #endif //DLIB_UTILITY_Hh_
